Fix add, drop, drag fucntionality
