+++ client/containers/PipelineContainer.tsx
@@ -1,0 +1,20 @@
+import { Container } from 'unstated';
+
+export interface PipelineContainerState {
+  currentPartner: string;
+}
+
+export default class PipelineContainer extends Container<
+  PipelineContainerState
+> {
+  constructor() {
+    super();
+    this.state = {
+      currentPartner: 'ALL',
+    };
+  }
+
+  setCurrentPartner(newPartner: string) {
+    this.setState({ currentPartner: newPartner });
+  }
+}
